Fix: 기타 버그 수정, 스킬 동작 테스트, 일부 스킬 범위 및 비용 조정
 - 대부분 기획안의 의도와 맞지 않게 작동하는 스킬들을 기획안대로 작동하게 조정했으며, 큰 변경은 없습니다.
 - 실드 로직을 실드 중첩시 스택이 쌓이게끔 변경하고, 실드가 버프 파괴효과에 파괴되지 않게 변경했습니다.
 - 테스트용으로 바꾼 마나 최대치를 원래대로 되돌렸습니다.
 - 스킬 툴팁에서, "분노" "비용"으로 표기된 것을 "마나"로 통일했습니다.
 - 일부 스킬 툴팁 통일성있게 변경

+++ Assets/0-Scripts/Objects/Characters/ExSkill/ExSkillLogics/SkillAbility_108.cs
@@ -30,7 +30,7 @@ public class SkillAbility_108 : ExSkillLogic
     public override void UseSkill(List<CharacterBase> target)// 스킬 실제 사용
     {
         //실드값 = 캐릭터 방어력 * (Ex스킬 계수 + (캐릭터 Ex스킬 레벨 * 캐릭터 Ex스킬 성장 계수)/100)
-        target[0].curCharacterBufList.AddBuf(BufKeyword.AmethystShield, 2, character, 0, (int)(character.Defend * (((float)(character.character.exSkill.coefficient[0].value) + (character.character.Growth.exSkillLevel * character.character.exSkill.growthCoefficient)) / 100)));
+        target[0].curCharacterBufList.AddBuf(BufKeyword.AmethystShield, 2, character, 0, (int)(character.Defend * (((float)(character.character.exSkill.coefficient[0].value) + (float)(character.character.Growth.exSkillLevel * character.character.exSkill.growthCoefficient)) / 100)));
     }
 
     public override void OnSkillAttackSuccess(CharacterBase target, BattleKeyWords.Damage damage)// 스킬 적중 시
